Add the wait queue data structure (linked list)
The wait queue defined by a linked list `wait_list`,
apart from ready queue, will store the processes in sleep state.

Since each entry stored in the queue (linked list) is
the address of a list_elem object in 'struct thread',
a new struct member (named waitelem) of 'struct thread' is introduced,
which has to be stored in the `wait_list` queue appropriately.