Implement thread sleeping (enqueueing wait queue)
When timer_sleep() is called, the thread is blocked by being enqueued
into the wait queue (rather than busy-waiting).

The function 'thread_sleep_until()' will mark the sleep_endtick
value defined in struct thread, the tick when the thread should be
awakend, and push the thread into the wait queue.

Then thread_block() will be called to switch to another thread.
To make thread_block() work correctly, we have to ensure that
interrupts should be turned off (in the timer routine).

In the alarm tests, threads are sleeping infinitely as of now,
since we didn't implement wake-up routines yet.