Basic implementation of acessing (reading) user memory
To retrieve the system call number, kernel should examine the stack
area of user process.

For the sake of it, functions `get_user` (reading a single byte) and
`memread_user` (reading a consecutive block of user memory) are
implemented. We also verify if a user pointer is valid area,
which would be the case the user program should be terminated.