Fix a memory leak, with regarding to wait() and exit()
There was a memory leak where pcb objects of exited processes
are not released, resulting the multi-oom tests to fail.

We have to ensure that a process should release all the resources
it was holding (file descriptors and its pcb object) on exiting.
The pcb object is, however, might be referenced by its parent
if the parent is waiting for the child process, so its deallocation
should be postponed until wait() is complete. Thus deallocation
of pcb object should be managed by its parent, either at the parent
is waiting for it or at the parent is terminating.

Now all tests passed, including multi-oom.