Fix resource management in case of orphan process
When a child process terminates after its parent does, it becomes
an orphan process. The parent can't free the pcb of the child yet,
since the child process is still alive.

So, we maintain a state 'orphan' for each process, so that the
parent would free only the child processes that had terminated before,
and that the orphan process can free itself when it terminates.