Implemented Stack Growth
Stack can now grow down (up to MAX_STACK_SIZE). In the page fault
handler, if the fault address lies on the stack frame, we install
a new virtual (zero) page on the supplemental page table, so that
the program won't die and the page on the stack could be accessed.

For the sake of it, vm_supt_install_zeropage() is newly implemented.
Moreover, we need to obtain the current value of stack pointer of
the user program, which is not trivial if the page fault occurs
in kernel mode; we store esp in the beginning of system calls so
that the page fault handler can lookup it. (as suggested in the manual)

All test cases related with stack growth now pass.