Implement basic swapping and eviction scheme
When allocation of new frame has failed (there is no space left),
some page should be evicted and swapped out into the swap disk.
An user page associated with the chosen frame is cleared and
updated to refer to the swap index (in the SUPT); just look at
vm_frame_allocate() for the details.

When a page fault has occured, where the page had been swapped out,
we have to swap it in, of course.