Free page-related resources on process termination
When process terminates, all the allocated frames and supplemental
page table entries associated to the process(thread) MUST be released.
It might happen that an another thread will try to access a frame
whose owner thread is already terminated, causing an invalid access.

To achieve this, we maintain a mapped kernel page (frame) within
the STPE, so that the frame entry could also be removed from the
frame table when the SUPT and its entries are destroyed on termination.