Pinning: pin pages prior to reading from or writing to filesys
When writing a buffer into the file, or when reading the file into
a buffer, if the buffer is not loaded (needs to be swapped in)
we get an unexpected page fault which results in a double-lock kernel
panic.

To prevent this, we pre-load and pin the frames, associated with the
target buffer, prior to do a file operation.