Maintain a list of frames for refining clock algorithm
Previous implementation of clock algorithm for evicting frames,
was rather naive because we manipulated only a hash table iterator.
To improve this, we also maintain a (conceptually circular) linked
list for the frame table entries, consisting of exactly same entries
of the frame hashmap. Then the clock algorithm implementation will
just utilize the list iterator `clock_ptr` which runs through the
frame linked list. The rest of strategy remains same.