Implement remaining system calls (readdir, isdir, inumber)
File descriptor of each process now contains a new field `dir`,
so that it can keep track of open directories. An open file is
considered to be a directory if `dir != NULL`. For the sake of this,
open() and close() system calls have to be changed to correctly
open and close the target directory.

Three remaining system calls, readdir(), isdir() and inumber() are
implemented. Also added related helper functions on inode layer.